fix: separate article paragraph stories into web and native
